Eliminate circular dependency
